Big heap of panel and layer work

+++ lib/Window.h
@@ -55,7 +55,7 @@ namespace OOGL
 		bool is_iconified() const;
 		void iconify(bool minimize);
 
-		glm::ivec2 size() const;
+		glm::u16vec2 size() const;
 		glm::vec2 dots_per_mm() const;
 
 		const OOBase::SharedPtr<Framebuffer>& get_default_frame_buffer() const;
@@ -65,8 +65,8 @@ namespace OOGL
 
 		OOBase::Delegate1<void,const Window&,OOBase::ThreadLocalAllocator> on_close(const OOBase::Delegate1<void,const Window&,OOBase::ThreadLocalAllocator>& delegate);
 		OOBase::Delegate2<void,const Window&,State&,OOBase::ThreadLocalAllocator> on_draw(const OOBase::Delegate2<void,const Window&,State&,OOBase::ThreadLocalAllocator>& delegate);
-		OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator> on_sized(const OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator>& delegate);
-		OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator> on_moved(const OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator>& delegate);
+		OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator> on_sized(const OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator>& delegate);
+		OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator> on_moved(const OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator>& delegate);
 		OOBase::Delegate3<void,const Window&,unsigned int,int,OOBase::ThreadLocalAllocator> on_character(const OOBase::Delegate3<void,const Window&,unsigned int,int,OOBase::ThreadLocalAllocator>& delegate);
 
 		struct key_stroke_t
@@ -87,8 +87,8 @@ namespace OOGL
 
 		OOBase::Delegate1<void,const Window&,OOBase::ThreadLocalAllocator> m_on_close;
 		OOBase::Delegate2<void,const Window&,State&,OOBase::ThreadLocalAllocator> m_on_draw;
-		OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator> m_on_sized;
-		OOBase::Delegate2<void,const Window&,const glm::ivec2&,OOBase::ThreadLocalAllocator> m_on_moved;
+		OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator> m_on_sized;
+		OOBase::Delegate2<void,const Window&,const glm::u16vec2&,OOBase::ThreadLocalAllocator> m_on_moved;
 		OOBase::Delegate3<void,const Window&,unsigned int,int,OOBase::ThreadLocalAllocator> m_on_character;
 		OOBase::Delegate2<void,const Window&,const key_stroke_t&,OOBase::ThreadLocalAllocator> m_on_keystroke;
